hr_agent/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
jinja2